refactor: Roulette으로 페이지 컴포넌트 네이밍 변경
